requirements.txt
altgraph==0.17.3
certifi==2022.12.7
chardet==3.0.4
customtkinter==5.1.2
darkdetect==0.8.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pywin32-ctypes==0.2.0
rfc3986==1.5.0
sniffio==1.3.0
